# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NyumbaLink API
  version: "1.0.0"
  description: |
    ## Kenya's Property Marketplace REST API

    **Base URL:** `https://nyumbalink.co.ke/api`

    ### Authentication
    All protected endpoints require a JWT Bearer token obtained from `/auth/login`.
    Include it as: `Authorization: Bearer <access_token>`

    ### Rate Limiting
    | Scope           | Limit             |
    |-----------------|-------------------|
    | Global          | 200 req/day, 50/hr |
    | Auth endpoints  | 10 req/min        |
    | File uploads    | 20 req/hr         |

    ### Pagination
    List endpoints accept:
    | Param     | Type    | Default | Description              |
    |-----------|---------|---------|--------------------------|
    | page      | integer | 1       | Page number (1-indexed)  |
    | per_page  | integer | 12      | Items per page (max 50)  |
    | sort_by   | string  | created_at | Column to sort by     |
    | sort_dir  | string  | desc    | `asc` or `desc`          |

  contact:
    name: NyumbaLink Support
    email: support@nyumbalink.co.ke
    url: https://nyumbalink.co.ke

  license:
    name: Proprietary
    url: https://nyumbalink.co.ke/terms
servers:
  - url: https://nyumbalink.co.ke/api
    description: Production
  - url: http://localhost:5000/api
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# ──────────────────────────────────────────────────────────────────────
#  Reusable schemas
# ──────────────────────────────────────────────────────────────────────
...
paths:
...
  /inquiries:
    post:
      summary: Send an inquiry about a property
      operationId: createInquiry
      tags: [Inquiries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [property_id, name, message]
              properties:
                property_id: { type: string, format: uuid }
                name:        { type: string }
                email:       { type: string, format: email }
                phone:       { type: string }
                message:     { type: string, minLength: 10 }
      responses:
        "201": { description: Inquiry sent, lead captured }
  /inquiries/my:
    get:
      summary: Get all inquiries for my properties (owner)
      operationId: myInquiries
      tags: [Inquiries]
      security:
        - BearerAuth: []
      parameters:
        - { name: page,     in: query, schema: { type: integer } }
        - { name: per_page, in: query, schema: { type: integer } }
      responses:
        "200": { description: Paginated inquiries }

  # ── Viewings ──────────────────────────────────────────────────────
  /viewings:
    post:
      summary: Request a property viewing
      operationId: createViewing
      tags: [Viewings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [property_id, preferred_date]
              properties:
                property_id:    { type: string, format: uuid }
                name:           { type: string }
                phone:          { type: string }
                email:          { type: string, format: email }
                preferred_date: { type: string, format: date }
                preferred_time: { type: string, example: "10:00" }
                message:        { type: string }
      responses:
        "201": { description: Viewing request created }
  /viewings/{id}/status:
    put:
      summary: Update viewing status (owner/admin)
      operationId: updateViewingStatus
      tags: [Viewings]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:         { type: string, enum: [confirmed, cancelled, completed] }
                confirmed_date: { type: string, format: date }
                confirmed_time: { type: string }
      responses:
        "200": { description: Status updated }

  # ── Payments ──────────────────────────────────────────────────────
...